+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Log/LoggerFiles.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GatoGraphQL\GatoGraphQL\Log;
+
+class LoggerFiles
+{
+    public const INFO = 'info.log';
+}